Refactor name of EditProfileActivity.kt to ProfileActivity.kt
