+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
             android:label="@string/title_activity_signup"
             android:theme="@style/Theme.CoachMe" />
         <activity
-            android:name=".EditProfileActivity"
+            android:name=".ProfileActivity"
             android:exported="false"
             android:label="@string/title_activity_edit_profile"
             android:theme="@style/Theme.CoachMe" />
