+++ app/src/main/java/com/github/sdpcoachme/DashboardActivity.kt
@@ -127,7 +127,7 @@ fun Dashboard(email: String, scaffoldState: ScaffoldState, onScaffoldStateChange
                 onItemClick = {
                     when (it.tag) {
                         PROFILE -> {
-                            val intent = Intent(context, EditProfileActivity::class.java)
+                            val intent = Intent(context, ProfileActivity::class.java)
                             intent.putExtra("email", email)
                             context.startActivity(intent)
                         }
